WordCamp Forms to Drafts: Restore a disabled version of form to avoid confusion
Hiding the form can confuse people because they expect to see a form and don't, making them wonder if they are on the wrong page, etc.

Showing a disabled version of it, along with making the warning message stand out more, makes it much more obvious that they need to perform an action before they can fill it out.

git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org@3877 74240141-8908-4e6f-9713-ba540dce6ec7