WordCamp Post Types: Fix sanitization of [speakers] params
In r5545 and r5546 some changes were made to the speakers shortcode
to enable a `track` and a `groups` parameter, both of which accept
a comma-separated list of term slugs. To sanitize these parameters,
the values went through an array map that trimmed whitespace. The
unintended side effect of this was that if the parameter wasn't
included in the shortcode, the sanitization would set the default value
to a non-empty array containing an empty string. This in turn caused
a `tax_query` parameter to be added to the WP_Query args, and the
shortcode would return an empty list for the speakers.

It turns out that trimming the whitespace isn't necessary for this case,
but removing empty array items is, so the `array_map` was swapped out
for `array_filter`. This seems to fix the empty speaker list issue, and
shortcodes using `track` or `groups` can include whitespace in their
comma-separated lists without a problem.

git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org@5547 74240141-8908-4e6f-9713-ba540dce6ec7