WordCamp: Temporarily disable the privacy tools in 4.9.6
At this point the privacy tools in Core are only designed for single site
installations. Using them on a large multisite network like WordCamp.org could
have negative unintended consequences, especially since the user table is
shared across all of the sites. Once we have a system in place for handling
privacy requests on a network-wide basis, we can re-enable the tools.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org@7193 74240141-8908-4e6f-9713-ba540dce6ec7