WordCamp: Avoid fatal error when Gutenberg isn't activated
`is_plugin_active` is usually scoped only for WP Admin, so it may not be
available when Gutenberg Tweaks is trying to determine if Gutenberg is
activated, which would throw a fatal error. Detecting the existence of a
function within the Gutenberg plugin itself is a safer option.

Fixes #3704
Props dryanpress



git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org@7459 74240141-8908-4e6f-9713-ba540dce6ec7