WordCamp Gutenberg Tweaks: Add hotfix for wpApiSettings bug
Starting in Gutenberg 3.1, the 'wp-api-request' script gets de-registered
and re-registered with a Gutenberg-specific version. Unfortunately, the
script doesn't also get re-localized with the `wpApiSettings` variable that's
needed for wp-api.js, unless the current screen contains an editor instance.

The bug was reported here:
https://github.com/WordPress/gutenberg/issues/7488

And it should be fixed once the changes in this pull request get released:
https://github.com/WordPress/gutenberg/pull/7566

In the meantime, this hotfix should ensure that `wpApiSettings` is available
regardless of the presence of an editor instance.



git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org@7476 74240141-8908-4e6f-9713-ba540dce6ec7