WordCamp Reports: Refactor WordCamp Counts for better data caching
This separates the data gathering so that `get_data` only handles the "raw"
rows of data that could be exported to a CSV, while `compile_report_data`
handles the counting of totals, uniques, etc. This also streamlines the
gender estimation process so that the API client is only used once per site.



git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org@7912 74240141-8908-4e6f-9713-ba540dce6ec7