WordCamp MES: Simplify saving of post meta to avoid accidental data loss
When an MES post is saved, the save routine logic looks for specific keys in
$_POST and for the ones it finds, it updates related post meta fields. For
ones it doesn't find, it deletes related post meta fields.

When updating an MES from the Edit screen, this is fine because all of the
postmeta values have related metabox fields, so they will always be present
in $_POST during the save_post action. But when using Quick Edit, those
metaboxes don't exist, so everything gets deleted.

Removing the lines that delete post meta fields when they are not present
fixes this issue. This should be ok because the fields will still be updated
with empty values if the related fields are cleared on the Edit screen.



git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org@8488 74240141-8908-4e6f-9713-ba540dce6ec7