WC Blocks: Refactor away deprecated `componentWillMount` function
We use `componentWillMount` and `componentWillUnmount` method in bunch of places. This method is soon to be deprecated, so it makes sense to move away from it.
As per best practices, we moved all `componentWillMount` method to inside constructor, since we were not subscribing to any events / not creating any side effects.

Fixes https://github.com/WordPress/wordcamp.org/issues/63

git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org@8600 74240141-8908-4e6f-9713-ba540dce6ec7