WC Camptix: Throttle ticket purchase request to prevent junk tickets.
This patch adds throttling support to ticket purchase request to 39 per hour. This is not intended to prevent any kind of sophiscated attacks, but is just there so that we don't have to delete lots of failed tickets when a security runs any automated tests on us.

Form_Spam_Prevention provides honeypot feature as well, which we are intentionally not using here because we want don't want to be aggressive in blocking ticket purchase form.

git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org@8636 74240141-8908-4e6f-9713-ba540dce6ec7