CampTix-Invoices: Hide the VAT table row if no VAT details are present.
Currently Invoices will include the VAT: 0 line in the table even if the plugin is set to not collect VAT numbers.

The attached patch simply hides that row, and only shows it if either of the following conditions are met:

1. The form is set to accept VAT numbers
2. The invoice has a VAT number associated with it

While using either of these conditionals individually would suffice, it was simpler to just cover future bases by checking both.

This will aid in the future usage of the plugin outside of the EU.

props dion

git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org@8797 74240141-8908-4e6f-9713-ba540dce6ec7