Open source drop-in plugins.
There's nothing sensitive in these, but previously they couldn't be open-sourced because SVN doesn't allow setting up an external to a single file, unless it's within the same repository.

Now that the Git repo is canonical, and we're simply syncing files to the production repository, there's no reason to keep these closed. Opening them will aid in collaboration, and keep development environments consistent.