JS Code Standards: Enforce a line-length of 115 characters (#150)
Set the `max-len` rule in eslint to 115 characters. The rule is configured to ignore line length in URLs, trailing comments, strings, and template literals. This prevents eslint from flagging "atomic" content, like a URL, that would be awkward to wrap.

Additionally, fix the few places where we were over 115 characters.