Tooling: Use Composer to manage third-party PHP packages
The impetus for doing this is so we can use third-party packages as
dependencies for some of our own custom modules and plugins. Also,
GitHub's recently added security features give us automatic
notifications when a dependency has a vulnerability and a new version
is available.

In addition to the `require` dependencies that add functionality
libraries to our codebase, all of the third-party plugins and themes
used on WordCamp.org are added as `require-dev` (development only)
dependencies. This replaces the previous bash script used to install
plugins and themes within the Docker development environment, and allows
us to specify the versions that get installed.

Some of the plugins/themes are installed via wpackagist, while others
have a custom package definition. The custom ones are such because they
either have a version reference that wpackagist doesn't support (i.e.
a branch instead of a tag), or the installation directory for the plugin
in our codebase doesn't match the plugin's slug in the WordPress plugin
directory.

**Note that running `composer install` (without `--no-dev`) should
never be done in the production environment, but only in local
development environments.**