Blocks: Use core-provided `@wordpress/scripts` build process and config.
This enables the `DependencyExtractionWebpackPlugin` plugin, so that `@wordpress/*` scripts are automatically extracted into a `*.deps.json` file, which we can read to automatically generate the correct dependencies for enqueuing. To get the benefit of dependency extraction, we need to import the wordpress components as `import`s,  rather than const declarations - the wp-scripts webpack config also takes care of converting those to the `wp.` globals, so we're not adding any extra libraries.

The switch brings in updated eslint & stylelint config, which flagged some errors. The following are fixed:

- The `BaseControl` component in `ImageSizeControl` & `ImageAlignmentControl` requires an `id` if a label is used. We don't want a `<label>` (semantically), so we can remove the label prop and use `<span>`s instead.
- In the scss files, remove the unit from `line-height` and remove a redundant `&`.

Fixes #71.