Blocks: Update `react-select` package.
Update react-select package to 3.0.4 to get improvements to performance & accessibility.

The update changed how the `onChange` callback is called, so now when removing the last selected item, `selectedOptions` is null. This causes a bug where `ContentSelect` defaults back to the passed in `value`, so instead we force an empty array when `selectedOptions` is falsy.