Service Worker Caching: Re-enable navigation preload.
Previously, it was necessary to disable it in order to get cached pages instead of the offline page. As of https://github.com/xwp/pwa-wp/pull/178, that is no longer the case, and it's better to enable it for improved performance.

That fix won't be available in production until a new version of the plugin is released and we update to it, but there aren't any sites currently using this, so it's better to get the change committed while everything is still fresh in my mind.