Blocks: Remove eslint overrides for declaration spacing. (#165)
As a WordPress community project, we should be following the core coding standards with as few modifications as possible, so that anyone involved in a core project can jump into our codebase without needing to learn a new style. Additionally, by overriding the `no-multi-spaces` rule, we're also opening the possibility of incorrect spacing elsewhere in the code.

Run `eslint --fix` to automatically fix spacing issues

Fixes #148, fixes #149.