Blocks: Load blocks on all sites (#213)
Remove the feature gate, to load new blocks on all WordCamp sites. Also add a check for which blocks should be loaded on a given page. This will allow us to feature gate specific blocks, by conditionally loading their controller (which populates `WordCampBlocks.[blockName]` with options and the schema). If the block is not set up in `WordCampBlocks`, we can assume it should not be loaded.