PWA: Cache custom & remote CSS (#212)
Move the CSS caching from precache to regular cache. The precache is a special cache storage that is longer lasting (and harder to update) than the regular cache. We'll use the regular cache for custom CSS instead of precaching this.

Add custom CSS and remote CSS endpoints to the routes cached using the "asset" group and strategy. This means custom and remote CSS will be cached for 1 day, and fetched from the cache first, with no network request if the resource is found.

This also bumps the number of entries allowed in the asset cache to 100, since a logged in user can see potentially 90+ assets loaded.