CampTix: Use CLDR data for country field values
This helps centralize the source of country name data into once place
within the WordCamp ecosystem, instead of having multiple hard-coded
lists of country names.

Note that this change does not need to be wrapped in a feature flag
conditional because the country field value is not validated against the
list of possible country names. So an existing attendee answer for the
country field that contains a country name string that is no longer in
the list will not be affected.