Blocks: Create a Live Schedule block (#243)
Add a new block called "Live Schedule", which can be used on the Day of Event template (or any other page). This copies the functionality from the Live Schedule component into a block. Eventually, this will let organizers have total control over the content of the Day of Event template.

Fixes #127 – You can customize the heading text and level in the block UI.
See #130 – Addresses a few tasks from this issue