Blocks: add "Live update" option to Latest Posts block (#250)
This will enable realtime updates of news posts on the frontend of the site, without refreshing the page. It can be used on the Day of Event template (or anywhere on the site) to always show the latest posts.

We hook into the Latest Posts block to add a new "Live Update" toggle, hook into the register_block_type code to unregister the core block, and re-register it with this new attribute and to add our front-end script. 

See #130