Subroles: Omit user caps that have been added via usermeta values (#355)
On WordCamp.org we grant additional capabilities to users via the map_meta_cap filter and a hard-coded list of users/caps. This way the granting/revoking of capabilities has an audit log in the form of revision history.

When a WP_User object is instantiated, it populates a caps property from a user meta value that is blog-specific. Normally this value only contains the user's role, but it can also store additional capabilities if they are added via add_cap. These additional capabilities are then merged with additions from map_meta_cap when evaluating something like current_user_can.

This adds a filter (and related unit tests!) that omits capabilities that have been incorporated from user meta when evaluating has_cap, so that only the map_meta_cap filter can add caps.