WordCamp Central: Add sponsors widget area to homepage (#314)
Replaces the Latest Tweets column on the WordCamp Central homepage with a Gold Sponsors widget area. The area will contain one Image Widget for each Gold-level sponsor in the Global Sponsorship program, which contains a logo image and an option caption tagline. The logos are shown one-by-one in a slideshow fashion.

Using a widget area makes it easier to adjust which logos are shown whenever the global sponsors program is updated, without having to change any code.

Fixes #306