Forms to Drafts: Fix login requirement on speaker application when using form block (#384)
Add the block delimiters to the check for form content on a post, so that the login overlay will take effect on block-forms as well as the shortcode. This brings back the wp.org login gateway on the call for speakers form, which should prevent misattributed speakers & submission spam. Additionally this adds the inert attribute (+ polyfill for browser support) to prevent keyboard users from being able to fill out & submit the form.

Fixes #332, Fixes #378