CampTix: Add "Remote" ticket type (#386)
Set up a "Ticket Type", which pulls from a global list of types (set up in code) and can conditionally alter the registration process. Types are set up as a safelist of slug+name, and the type is saved in ticket post meta. This also updates the allergy & accommodations questions, by adding filters to skip rendering the question and/or reword the question text. Other ticket types are possible in the future, by filtering `camptix_ticket_types` & similarly working with CampTix filters.

Fixes #385