Speaker Feedback: Check submissions for spam, duplicates (#406)
Our feedback submissions require a custom configuration for Akismet to check them correctly for spam, even though they are comments. This commit ensures that submissions get sent successfully to the Akismet API when processing requests to our REST API endpoint. It also ensures that if a feedback comment is later manually set to spam or not spam, the correct info is sent to the Akismet API so it can "learn". All of this happens through our custom list table UI such that the spam workflow is identical to that of normal comments.

Finally, this fixes a bug that was allowing duplicate feedback comments to successfully be submitted.

On the back end, this is using the Akismet plugin and much as possible, and Jetpack's Contact Form module for some things as well (namely sending the actual request to the Akismet API). These are both plugins that are available on every WordCamp site, and this allows us to avoid reinventing the wheel on a lot of functionality and UI for spam management.

Fixes #346